+++ src/splitter_mr/reader/base_reader.py
@@ -155,4 +155,3 @@ class BaseReader(ABC):
                     )
             ```
         """
-        pass